lab/opengl/requirements.txt
pyOpenGL
numpy
glfw
PyGLM